+++ nilearn/plotting/html_surface.py
@@ -95,6 +95,7 @@ def full_brain_info(volume_img, mesh='fsaverage5', threshold=None,
     mesh = surface.surface._check_mesh(mesh)
     surface_maps = {
         h: surface.vol_to_surf(volume_img, mesh['pial_{}'.format(h)],
+                               inner_mesh=mesh.get('white_{}'.format(h), None),
                                **vol_to_surf_kwargs)
         for h in ['left', 'right']
     }